Creating an interface for error mail
